connector*: fix index name and prefix name stability
Having index_prefix_name coming from env is a must.
Originally we were relying on the fact that server_env would update the value
at each restart based on the env value.

This is not working anymore in v18 and we tried already to fix this.

One of the other goals (still possible up to v14) was to have one single config entry
only for the specific backend implementation, not to have to deal with 2 different configurations.

This as well was not working properly in particular because the index name was stored
hence a forced recompute was needed but also was causing caching issues on views.

In the end the solution is to keep it simple:

* index name is not stored anymore
* se.backend and se.backend.*specific* will need 2 different configurations